+++ src/readable/byob_reader.rs
@@ -1,7 +1,7 @@
 use std::marker::PhantomData;
 
-use js_sys::Uint8Array;
-use wasm_bindgen::{throw_val, JsCast, JsValue};
+use js_sys::{Object, Uint8Array};
+use wasm_bindgen::{JsCast, JsValue};
 use wasm_bindgen_futures::JsFuture;
 
 use crate::util::{checked_cast_to_usize, clamp_to_u32, promise_to_void_future};
@@ -23,9 +23,14 @@ pub struct ReadableStreamBYOBReader<'stream> {
 impl<'stream> ReadableStreamBYOBReader<'stream> {
     pub(crate) fn new(stream: &mut ReadableStream) -> Result<Self, js_sys::Error> {
         Ok(Self {
-            raw: stream.as_raw().get_reader_with_options(
-                sys::ReadableStreamGetReaderOptions::new(sys::ReadableStreamReaderMode::BYOB),
-            )?,
+            raw: stream
+                .as_raw()
+                .unchecked_ref::<sys::ReadableStreamExt>()
+                .try_get_reader_with_options(
+                    sys::ReadableStreamGetReaderOptions::new()
+                        .mode(sys::ReadableStreamReaderMode::Byob),
+                )?
+                .unchecked_into(),
             _stream: PhantomData,
         })
     }
@@ -109,20 +114,18 @@ impl<'stream> ReadableStreamBYOBReader<'stream> {
         let buffer_len = buffer.byte_length();
         // Limit view to destination slice's length.
         let dst_len = clamp_to_u32(dst.len());
-        let view = buffer
-            .subarray(0, dst_len)
-            .unchecked_into::<sys::ArrayBufferView>();
+        let view = buffer.subarray(0, dst_len).unchecked_into::<Object>();
         // Read into view. This transfers `buffer.buffer()`.
-        let promise = self.as_raw().read(&view);
-        let js_value = JsFuture::from(promise).await?;
-        let result = sys::ReadableStreamBYOBReadResult::from(js_value);
-        let filled_view = match result.value() {
-            Some(view) => view,
-            None => {
-                // No new view was returned. The stream must have been canceled.
-                assert!(result.is_done());
-                return Ok((0, None));
-            }
+        let promise = self.as_raw().read_with_array_buffer_view(&view);
+        let js_result = JsFuture::from(promise).await?;
+        let result = sys::ReadableStreamReadResult::from(js_result);
+        let js_value = result.value();
+        let filled_view = if js_value.is_undefined() {
+            // No new view was returned. The stream must have been canceled.
+            assert!(result.is_done());
+            return Ok((0, None));
+        } else {
+            js_value.unchecked_into::<Uint8Array>()
         };
         let filled_len = checked_cast_to_usize(filled_view.byte_length());
         debug_assert!(filled_len <= dst.len());
@@ -157,9 +160,7 @@ impl<'stream> ReadableStreamBYOBReader<'stream> {
     }
 
     fn release_lock_mut(&mut self) {
-        self.as_raw()
-            .release_lock()
-            .unwrap_or_else(|error| throw_val(error.into()))
+        self.as_raw().release_lock()
     }
 
     /// Try to [release](https://streams.spec.whatwg.org/#release-a-lock) this reader's lock on the
@@ -175,7 +176,10 @@ impl<'stream> ReadableStreamBYOBReader<'stream> {
     /// return an error and leave the reader locked to the stream.
     #[inline]
     pub fn try_release_lock(self) -> Result<(), (js_sys::Error, Self)> {
-        self.as_raw().release_lock().map_err(|error| (error, self))
+        self.as_raw()
+            .unchecked_ref::<sys::ReadableStreamReaderExt>()
+            .try_release_lock()
+            .map_err(|err| (err, self))
     }
 
     /// Converts this `ReadableStreamBYOBReader` into an [`AsyncRead`].
@@ -185,7 +189,7 @@ impl<'stream> ReadableStreamBYOBReader<'stream> {
     /// still usable. This allows reading only a few bytes from the `AsyncRead`, while still
     /// allowing another reader to read the remaining bytes later on.
     ///
-    /// [`AsyncRead`]: https://docs.rs/futures/0.3.18/futures/io/trait.AsyncRead.html
+    /// [`AsyncRead`]: https://docs.rs/futures/0.3.28/futures/io/trait.AsyncRead.html
     #[inline]
     pub fn into_async_read(self) -> IntoAsyncRead<'stream> {
         IntoAsyncRead::new(self, false)