+++ src/readable/into_async_read.rs
@@ -4,14 +4,14 @@ use core::task::{Context, Poll};
 use futures_util::io::{AsyncRead, Error};
 use futures_util::ready;
 use futures_util::FutureExt;
-use js_sys::Uint8Array;
+use js_sys::{Object, Uint8Array};
 use wasm_bindgen::prelude::*;
 use wasm_bindgen::JsCast;
 use wasm_bindgen_futures::JsFuture;
 
 use crate::util::{checked_cast_to_usize, clamp_to_u32, js_to_io_error};
 
-use super::sys::{ArrayBufferView, ReadableStreamBYOBReadResult};
+use super::sys::ReadableStreamReadResult;
 use super::ReadableStreamBYOBReader;
 
 /// An [`AsyncRead`] for the [`into_async_read`](super::ReadableStream::into_async_read) method.
@@ -20,7 +20,7 @@ use super::ReadableStreamBYOBReader;
 /// When this `AsyncRead` is dropped, it also drops its reader which in turn
 /// [releases its lock](https://streams.spec.whatwg.org/#release-a-lock).
 ///
-/// [`AsyncRead`]: https://docs.rs/futures/0.3.18/futures/io/trait.AsyncRead.html
+/// [`AsyncRead`]: https://docs.rs/futures/0.3.28/futures/io/trait.AsyncRead.html
 #[must_use = "readers do nothing unless polled"]
 #[derive(Debug)]
 pub struct IntoAsyncRead<'reader> {
@@ -84,13 +84,12 @@ impl<'reader> AsyncRead for IntoAsyncRead<'reader> {
                     _ => Uint8Array::new_with_length(buf_len),
                 };
                 // Limit to output buffer size
-                let buffer = buffer
-                    .subarray(0, buf_len)
-                    .unchecked_into::<ArrayBufferView>();
+                let buffer = buffer.subarray(0, buf_len).unchecked_into::<Object>();
                 match &self.reader {
                     Some(reader) => {
                         // Read into internal buffer and store its future
-                        let fut = JsFuture::from(reader.as_raw().read(&buffer));
+                        let fut =
+                            JsFuture::from(reader.as_raw().read_with_array_buffer_view(&buffer));
                         self.fut.insert(fut)
                     }
                     None => {
@@ -108,14 +107,14 @@ impl<'reader> AsyncRead for IntoAsyncRead<'reader> {
         // Read completed
         Poll::Ready(match js_result {
             Ok(js_value) => {
-                let result = ReadableStreamBYOBReadResult::from(js_value);
+                let result = ReadableStreamReadResult::from(js_value);
                 if result.is_done() {
                     // End of stream
                     self.discard_reader();
                     Ok(0)
                 } else {
                     // Cannot be canceled, so view must exist
-                    let filled_view = result.value().unwrap_throw();
+                    let filled_view = result.value().unchecked_into::<Uint8Array>();
                     // Copy bytes to output buffer
                     let filled_len = checked_cast_to_usize(filled_view.byte_length());
                     debug_assert!(filled_len <= buf.len());