+++ src/readable/into_underlying_source.rs
@@ -82,7 +82,7 @@ impl Inner {
         // after the stream has closed or encountered an error.
         let stream = self.stream.as_mut().unwrap_throw();
         match stream.try_next().await {
-            Ok(Some(chunk)) => controller.enqueue(&chunk)?,
+            Ok(Some(chunk)) => controller.enqueue_with_chunk(&chunk)?,
             Ok(None) => {
                 // The stream has closed, drop it.
                 self.stream = None;