+++ src/writable/mod.rs
@@ -25,7 +25,7 @@ pub mod sys;
 /// They can be converted into a [raw JavaScript stream](sys::WritableStream) with
 /// [`into_raw`](Self::into_raw), or into a Rust [`Sink`] with [`into_sink`](Self::into_sink).
 ///
-/// [`Sink`]: https://docs.rs/futures/0.3.18/futures/sink/trait.Sink.html
+/// [`Sink`]: https://docs.rs/futures/0.3.28/futures/sink/trait.Sink.html
 #[derive(Debug)]
 pub struct WritableStream {
     raw: sys::WritableStream,
@@ -44,9 +44,9 @@ impl WritableStream {
     /// Use [`with`] and/or [`sink_map_err`] to convert a sink's items to a `JsValue`
     /// before passing it to this function.
     ///
-    /// [`Sink`]: https://docs.rs/futures/0.3.18/futures/sink/trait.Sink.html
-    /// [`with`]: https://docs.rs/futures/0.3.18/futures/sink/trait.SinkExt.html#method.with
-    /// [`sink_map_err`]: https://docs.rs/futures/0.3.18/futures/sink/trait.SinkExt.html#method.sink_map_err
+    /// [`Sink`]: https://docs.rs/futures/0.3.28/futures/sink/trait.Sink.html
+    /// [`with`]: https://docs.rs/futures/0.3.28/futures/sink/trait.SinkExt.html#method.with
+    /// [`sink_map_err`]: https://docs.rs/futures/0.3.28/futures/sink/trait.SinkExt.html#method.sink_map_err
     pub fn from_sink<Si>(sink: Si) -> Self
     where
         Si: Sink<JsValue, Error = JsValue> + 'static,
@@ -54,7 +54,7 @@ impl WritableStream {
         let sink = IntoUnderlyingSink::new(Box::new(sink));
         // Use the default queuing strategy (with a HWM of 1 chunk).
         // We shouldn't set HWM to 0, since that would break piping to the writable stream.
-        let raw = sys::WritableStream::new_with_sink(sink);
+        let raw = sys::WritableStreamExt::new_with_into_underlying_sink(sink).unchecked_into();
         WritableStream { raw }
     }
 
@@ -73,7 +73,7 @@ impl WritableStream {
     /// Returns `true` if the stream is [locked to a writer](https://streams.spec.whatwg.org/#lock).
     #[inline]
     pub fn is_locked(&self) -> bool {
-        self.as_raw().is_locked()
+        self.as_raw().locked()
     }
 
     /// [Aborts](https://streams.spec.whatwg.org/#abort-a-writable-stream) the stream,
@@ -126,9 +126,9 @@ impl WritableStream {
     /// **Panics** if the stream is already locked to a writer. For a non-panicking variant,
     /// use [`try_into_sink`](Self::try_into_sink).
     ///
-    /// [`Sink`]: https://docs.rs/futures/0.3.18/futures/sink/trait.Sink.html
-    /// [`with`]: https://docs.rs/futures/0.3.18/futures/sink/trait.SinkExt.html#method.with
-    /// [`sink_map_err`]: https://docs.rs/futures/0.3.18/futures/sink/trait.SinkExt.html#method.sink_map_err
+    /// [`Sink`]: https://docs.rs/futures/0.3.28/futures/sink/trait.Sink.html
+    /// [`with`]: https://docs.rs/futures/0.3.28/futures/sink/trait.SinkExt.html#method.with
+    /// [`sink_map_err`]: https://docs.rs/futures/0.3.28/futures/sink/trait.SinkExt.html#method.sink_map_err
     #[inline]
     pub fn into_sink(self) -> IntoSink<'static> {
         self.try_into_sink()
@@ -144,9 +144,9 @@ impl WritableStream {
     /// If the stream is already locked to a writer, then this returns an error
     /// along with the original `WritableStream`.
     ///
-    /// [`Sink`]: https://docs.rs/futures/0.3.18/futures/sink/trait.Sink.html
-    /// [`with`]: https://docs.rs/futures/0.3.18/futures/sink/trait.SinkExt.html#method.with
-    /// [`sink_map_err`]: https://docs.rs/futures/0.3.18/futures/sink/trait.SinkExt.html#method.sink_map_err
+    /// [`Sink`]: https://docs.rs/futures/0.3.28/futures/sink/trait.Sink.html
+    /// [`with`]: https://docs.rs/futures/0.3.28/futures/sink/trait.SinkExt.html#method.with
+    /// [`sink_map_err`]: https://docs.rs/futures/0.3.28/futures/sink/trait.SinkExt.html#method.sink_map_err
     pub fn try_into_sink(mut self) -> Result<IntoSink<'static>, (js_sys::Error, Self)> {
         let writer = WritableStreamDefaultWriter::new(&mut self).map_err(|err| (err, self))?;
         Ok(writer.into_sink())
@@ -159,7 +159,7 @@ impl WritableStream {
     /// **Panics** if the stream is already locked to a writer. For a non-panicking variant,
     /// use [`try_into_async_write`](Self::try_into_async_write).
     ///
-    /// [`AsyncWrite`]: https://docs.rs/futures/0.3.18/futures/io/trait.AsyncWrite.html
+    /// [`AsyncWrite`]: https://docs.rs/futures/0.3.28/futures/io/trait.AsyncWrite.html
     pub fn into_async_write(self) -> IntoAsyncWrite<'static> {
         self.try_into_async_write()
             .expect_throw("already locked to a writer")
@@ -172,7 +172,7 @@ impl WritableStream {
     /// If the stream is already locked to a writer, then this returns an error
     /// along with the original `WritableStream`.
     ///
-    /// [`AsyncWrite`]: https://docs.rs/futures/0.3.18/futures/io/trait.AsyncWrite.html
+    /// [`AsyncWrite`]: https://docs.rs/futures/0.3.28/futures/io/trait.AsyncWrite.html
     pub fn try_into_async_write(self) -> Result<IntoAsyncWrite<'static>, (js_sys::Error, Self)> {
         Ok(IntoAsyncWrite::new(self.try_into_sink()?))
     }