+++ tests/tests/mod.rs
@@ -1,3 +1,4 @@
+mod fetch_as_stream;
 mod pipe;
 mod readable_byte_stream;
 mod readable_stream;